Add optional user callback UserBulkInitHook
The new callback `UserBulkInitHook` is optionally assigned by the user
using the `setUserBulkInitHook` method on the `FlatfileImporter` class.
If assigned, this hook is invoked on all rows passed to the
`bulkHookCallback` shorter after function entry.

The idea behind the bulk init hook is to support user workflows where
processing the entire dataset in larger chunks is more efficient than
processing individual rows at a time.